finalizing initial TODOs
