+++ src/fieldToConceptIdMapping.js
@@ -54,7 +54,7 @@ export default {
     no: 104430631,
     
     // collection id
-    collectionId: 825582494, //TODO this should be 'objectId' per the data dictionary
+    collectionId: 825582494,
     dateReceived: 926457119,
     dateWithdrawn: 678166505,
     collectionType: 650516960,
