+++ src/pages/checkOutReport.js
@@ -63,14 +63,14 @@ export const populateCheckOutTable = async () => {
     
     showAnimation();
     const data = await findParticipant(`checkedIn=${true}`).then(res => res.data);
-    
+    console.log('participant data', data);
     for (const item of data) {
       if (!item[fieldToConceptIdMapping.collection.selectedVisit]?.[fieldToConceptIdMapping.baseline.visitId]?.[fieldToConceptIdMapping.checkOutDateTime]) {
         const newRow = currTable.insertRow();
         newRow.innerHTML = `
           <td>${item['Connect_ID']}</td>
-          <td>${item[fieldToConceptIdMapping.lName]}</td>
-          <td>${item[fieldToConceptIdMapping.fName]}</td>
+          <td>${item[fieldToConceptIdMapping.lastName]}</td>
+          <td>${item[fieldToConceptIdMapping.firstName]}</td>
           <td>${convertISODateTime(item[fieldToConceptIdMapping.collection.selectedVisit]?.[fieldToConceptIdMapping.baseline.visitId]?.[fieldToConceptIdMapping.checkInDateTime])}</td>
           <td><button class="btn btn-outline-primary text-nowrap participantCheckOutBtn" data-checkout='${JSON.stringify(item)}'>Go to Check-Out</button></td>
         `;
