+++ src/pages/shipping.js
@@ -1,14 +1,15 @@
-import { appState, userAuthorization, removeActiveClass, displayContactInformation, getBoxes, getAllBoxes, hideAnimation, showAnimation, showNotifications, locationConceptIDToLocationMap } from "./../shared.js"
-import { addEventBackToSearch, addEventAddSpecimenToBox, populateAvailableCollectionsList, addEventNavBarShipment, addEventNavBarBoxManifest, formatBoxTimestamp, populateBoxManifestTable, populateBoxesToShipTable,
-         populateShippingManifestBody,populateShippingManifestHeader, addEventNavBarShippingManifest, populateTrackingQuery, addEventCompleteButton, populateFinalCheck, populateViewShippingBoxContentsList, addEventBoxSelectListChanged,
-         addEventCompleteShippingButton, populateSelectLocationList, addEventModalAddBox, populateTempNotification, populateTempCheck, populateTempSelect, addEventNavBarTracking, addEventReturnToReviewShipmentContents,
-         populateCourierBox, addEventSaveButton, addEventTrimTrackingNums, addEventCheckValidTrackInputs, addEventPreventTrackingConfirmPaste, addEventReturnToPackaging, addEventShipPrintManifest, addEventTrackingNumberScanAutoFocus } from "./../events.js";
+import { addBox, appState, conceptIdToSiteSpecificLocation, displayContactInformation, getAllBoxes, getLocationsInstitute, hideAnimation, locationConceptIDToLocationMap,
+        removeActiveClass, removeBag, removeMissingSpecimen, showAnimation, showNotifications, siteSpecificLocation, siteSpecificLocationToConceptId, sortBiospecimensList,
+        translateNumToType, userAuthorization } from "./../shared.js"
+import { addDeviationTypeCommentsContent, addEventAddSpecimenToBox, addEventBackToSearch, addEventBoxSelectListChanged, addEventCheckValidTrackInputs, addEventCompleteButton,
+        addEventCompleteShippingButton, addEventModalAddBox, addEventNavBarBoxManifest, addEventNavBarShipment, addEventNavBarShippingManifest, addEventNavBarTracking, addEventLocationSelect,
+        addEventPreventTrackingConfirmPaste, addEventReturnToPackaging, addEventReturnToReviewShipmentContents, addEventSaveButton, addEventShipPrintManifest,
+        addEventTrackingNumberScanAutoFocus, addEventTrimTrackingNums, compareBoxIds, getInstituteSpecimensList, populateCourierBox, populateFinalCheck, populateTrackingQuery } from "./../events.js";     
 import { homeNavBar, shippingNavBar, unAuthorizedUser} from '../navbar.js';
-import { setAllShippingState } from "../shippingState.js";
+import { setAllShippingState, updateShippingStateCreateBox, updateShippingStateRemoveBagFromBox } from '../shippingState.js';
 import conceptIds from '../fieldToConceptIdMapping.js';
 
 export const shippingDashboard = (auth, route) => {
-    console.log("calling shippingDashboard");  
     auth.onAuthStateChanged(async user => {
         if (user) {
             console.time('userAuthorization');
@@ -30,10 +31,8 @@ export const shippingDashboard = (auth, route) => {
 }
 
 
-// Initialize the shipping page.
-// Check for a stored location, else wait for a location to be selected, then build the page
+// Check for a stored location and initialize the shipping page.
 export const startShipping = async (userName, loadFromState = false, currBoxId) => {    
-    console.log('START SHIPPING', currBoxId);
     buildNavAndHeaderDOM();
 
     const storedLocation = getStoredLocationOnInit();
@@ -64,14 +63,14 @@ const buildShippingDOM = async () => {
 }
 
 /**
- * NOTE: getAllBoxesList is used because the larger filter iterates ALL boxes and removes specimens that belong to the shipped boxes ¯\_(ツ)_/¯
+ * NOTE: allBoxesList is used because the larger filter iterates ALL boxes and removes specimens that belong to the shipped boxes.
+ * TODO: future change: add shipped property and only pull unshipped.
  * This will run asyncronously when pulling new data (loadFromState = false) and synchronously when loading from state (loadFromState = true).
  * @param {string} selectedLocation - the user's selected location from local storage.
  * @param {*} userName - the logged-in userName.
  * @param {*} loadFromState - whether to load data from state or from the server.
  */
 const buildShippingInterface = async (selectedLocation, userName, loadFromState, currBoxId) => {    
-    console.log('calling - BUILD SHIPPING INTERFACE');
     showAnimation();
 
     let allBoxesList;
@@ -103,14 +102,12 @@ const buildShippingInterface = async (selectedLocation, userName, loadFromState,
     hideAnimation();
 }
 
-//TODO: use state for userName refs
 const addShippingEventListeners = () => {
     const userName = appState.getState().userName;
-    const tempMonitorCheckedEl = document.getElementById('tempMonitorChecked');
     addEventNavBarShipment("navBarShippingDash", userName);
-    addEventNavBarShippingManifest(userName, tempMonitorCheckedEl);
+    addEventNavBarShippingManifest(userName);
     addEventBoxSelectListChanged();
-    addEventNavBarBoxManifest("navBarBoxManifest", userName); //TODO does this get used?
+    addEventNavBarBoxManifest("navBarBoxManifest");
     addEventLocationSelect("selectLocationList", "shipping_location");
     addEventAddSpecimenToBox();
     addEventModalAddBox();
@@ -121,28 +118,893 @@ const getStoredLocationOnInit = () => {
 }
 
 /**
- * Location selection event listener. Remove the event listener if it was previously added, then add the event listener back.
- * This makes sure the listener is only added once.
- * @param {string} elemId - the id of the element to add the event listener to 
- * @param {*} pageAndElement - the page and element to store in local storage
- * @param {*} userName - the user name to store in local storage
+ * Populate the 'Available Collections' table.
+ * @param {array} boxList - list of available boxes.
+ * @param {boolean} loadFromState - if true, load data from state instead of fetching from server.
+ * @returns {array} finalizedSpecimenList - the list of specimens with the finalized flag.
+ * @returns {object} availableCollectionsObj - obj to populate the 'Available Collections' table.
+ * If finalizedSpecimenList or availableCollectionsObj are empty (such as initial load), fetched from server.
+ * Note: Orphan panel is currently hidden by request of the product team. Retain for future use.
+ *       Future orphan panel use would require completed state management implementation in the 'currDeleteButton' event listener.
  */
-const addEventLocationSelect = (elemId, pageAndElement) => {
-    const selectionChangeHandler = (event) => {
-        const selection = event.target.value;
-        const prevSelections = JSON.parse(localStorage.getItem('selections'));
-        localStorage.setItem('selections', JSON.stringify({...prevSelections, [pageAndElement] : selection}));
-        if (selection && selection !== 'none') {
-            const currBoxId = document.getElementById('selectBoxList').value;
-            startShipping(appState.getState().userName, true, currBoxId);
+const populateAvailableCollectionsList = async (boxList, loadFromState = false) => {
+    let finalizedSpecimenList = appState.getState().finalizedSpecimenList ?? [];
+    let availableCollectionsObj = appState.getState().availableCollectionsObj ?? {};
+
+    if (!loadFromState) {
+        ({ finalizedSpecimenList, availableCollectionsObj } = await getInstituteSpecimensList(boxList));
+    }
+
+    const bagIdList = Object.keys(availableCollectionsObj).sort();
+
+    const tableEle = document.getElementById("specimenList");
+    tableEle.innerHTML = buildPopulateSpecimensHeader();
+    
+    let numRows = 1;
+    let orphanBagId;
+
+    for (const bagId of bagIdList) {
+        if (bagId != "unlabelled") {
+            const rowEle = tableEle.insertRow();
+            rowEle.insertCell(0).innerHTML = bagId;
+            rowEle.insertCell(1).innerHTML = availableCollectionsObj[bagId].length;
+
+            const hiddenChannel = rowEle.insertCell(2)
+            hiddenChannel.innerHTML = JSON.stringify(availableCollectionsObj[bagId]);
+            hiddenChannel.style.display = "none";
+            if (numRows % 2 == 0) {
+                rowEle.style['background-color'] = "lightgrey";
+            }
+            numRows += 1;
+        } else {
+            orphanBagId = bagId;
         }
+    }
+
+    const orphanPanel = document.getElementById('orphansPanel');
+    const orphanTableEle = document.getElementById('orphansList');
+    const specimenPanel = document.getElementById('specimenPanel');
+    orphanTableEle.innerHTML = '';
+
+    if (orphanBagId && availableCollectionsObj['unlabelled'].length > 0) {
+        orphanPanel.style.display = 'block';
+        specimenPanel.style.height = '550px';
+
+        const orphanTubeIdList = availableCollectionsObj['unlabelled'];
+        const rowEle = orphanTableEle.insertRow();
+        rowEle.insertCell(0).innerHTML = 'Stray tubes';
+        rowEle.insertCell(1).innerHTML = orphanTubeIdList.length;
+        const hiddenChannel = rowEle.insertCell(2);
+        hiddenChannel.innerHTML = JSON.stringify(orphanTubeIdList);
+        hiddenChannel.style.display = "none";
+
+        for (let i = 0; i < orphanTubeIdList.length; i++) {
+            const rowCount = orphanTableEle.rows.length;
+            const rowEle = orphanTableEle.insertRow();
+
+            if (rowCount % 2 == 0) {
+                rowEle.style['background-color'] = 'lightgrey';
+            }
+
+            rowEle.insertCell(0).innerHTML = orphanTubeIdList[i];
+            rowEle.insertCell(1).innerHTML = '<input type="button" class="delButton" value = "Report as Missing"/>';
+
+            const currDeleteButton = rowEle.cells[1].getElementsByClassName("delButton")[0];
+
+            //This should remove the entrire bag
+            currDeleteButton.addEventListener("click", async e => {
+                showAnimation();
+                const index = e.target.parentNode.parentNode.rowIndex;
+                const table = e.target.parentNode.parentNode.parentNode.parentNode;
+                const currTubeId = table.rows[index].cells[0].innerText;
+                await removeMissingSpecimen(currTubeId);
+                hideAnimation();
+
+                startShipping(appState.getState().userName);
+            });
+        }
+    } else {
+        orphanPanel.style.display = 'none'
+        specimenPanel.style.height = '550px'
+    }
+
+    return { finalizedSpecimenList, availableCollectionsObj };
+}
+
+const populateBoxesToShipTable = () => {
+    const detailedBoxes = appState.getState().detailedProviderBoxes;
+    const table = document.getElementById("saveTable");
+    table.innerHTML = renderBoxesToShipTableHeader();
+
+    if (Object.keys(detailedBoxes).length > 0) {
+        const sortedBoxKeys = Object.keys(detailedBoxes).sort();
+        sortedBoxKeys.forEach((box, i) => {
+            const currBox = detailedBoxes[box];
+            const bagKeys = Object.keys(currBox).filter(key => key !== 'boxData' && key !== 'undefined').sort((a, b) => a.split(/\s+/)[1] - b.split(/\s+/)[1]);
+            const boxStartedTimestamp = currBox.boxData[conceptIds.firstBagAddedToBoxTimestamp] ? formatTimestamp(Date.parse(currBox.boxData[conceptIds.firstBagAddedToBoxTimestamp])) : '';
+            const boxLastModifiedTimestamp = currBox.boxData[conceptIds.shippingShipDateModify] ? formatTimestamp(Date.parse(currBox.boxData[conceptIds.shippingShipDateModify])) : '';
+            const boxLocation = currBox.boxData[conceptIds.shippingLocation] ? locationConceptIDToLocationMap[currBox.boxData[conceptIds.shippingLocation]]["siteSpecificLocation"] : '';
+            const numTubesInBox = bagKeys.reduce((total, bagKey) => total + currBox[bagKey]['arrElements'].length, 0);
+
+            const currRow = table.insertRow(i + 1);
+            currRow.style['background-color'] = i % 2 === 1 ? 'lightgrey' : '';
+            currRow.innerHTML += renderBoxesToShipRow(boxStartedTimestamp, boxLastModifiedTimestamp, box, boxLocation, numTubesInBox);
+            
+            const currBoxButton = currRow.cells[6].querySelector(".boxManifestButton");
+            currBoxButton.addEventListener("click", async () => {
+                generateBoxManifest(currBox);
+            });
+        });
+    }
+}
+
+export const populateTempNotification = () => {
+
+    let checkDate = false;
+    let toToggle = document.getElementById('tempTubeReminder');
+    if (checkDate == true) {
+        toToggle.style.display = 'block';
+    }
+    else {
+        toToggle.style.display = 'none';
+    }
+}
+
+export const populateTempCheck = () => {
+    let checkDate = false;
+    let toToggle = document.getElementById('checkForTemp');
+    if (checkDate == true) {
+        toToggle.style.display = 'block';
+    }
+    else {
+        toToggle.style.display = 'none';
+    }
+}
+
+const populateShippingBoxContentsRows = (bagNum, cellNum, row, currBagId, fullTubeId, tubeType) => {
+    bagNum % 2 == 1 ? row.style['background-color'] = 'lightgrey' : row.style['background-color'] = 'white'; {
+        row.insertCell(0).innerHTML = (cellNum === 0) ? currBagId : '';
+        row.insertCell(1).innerHTML = fullTubeId;
+        row.insertCell(2).innerHTML = tubeType;
+        row.insertCell(3).innerHTML = (cellNum === 0) ? '<input type="button" class="delButton" value = "remove bag" style="margin-top:2px;margin-bottom:2px">' : '';
+    }
+}
+
+const handleShippingBoxContentsSelector = (boxIdAndBagsObj, selectedBoxId) => {
+    const boxIdArray = Object.keys(boxIdAndBagsObj).sort(compareBoxIds);
+    const boxOptionsList = boxIdArray.map(boxId => `<option>${boxId}</option>`).join('');
+    const boxSelectEle = document.getElementById('selectBoxList');
+    boxSelectEle.innerHTML = boxOptionsList;
+    boxSelectEle.value = selectedBoxId ?? boxIdArray[0];
+
+    return boxSelectEle.value;
+}
+
+// This is the list on shipping screen -> view shipping box contents
+export const populateViewShippingBoxContentsList = (selectedBoxId) => {
+    const boxIdAndBagsObj = appState.getState().detailedProviderBoxes;
+    const currBoxId = handleShippingBoxContentsSelector(boxIdAndBagsObj, selectedBoxId);
+    
+    if (currBoxId != '') {
+        const currBox = boxIdAndBagsObj[currBoxId];
+        const shippingBoxContentsTable = document.getElementById('currTubeTable');
+        const boxKeys = Object.keys(currBox).filter(key => key !== 'boxData' && key !== 'undefined');
+
+        shippingBoxContentsTable.innerHTML = renderViewBoxContentsTableHeader();
+
+        //set up the table
+        for (let bagNum = 0; bagNum < boxKeys.length; bagNum++) {
+            const currBagId = boxKeys[bagNum];
+            const currTubes = currBox[boxKeys[bagNum]]['arrElements'];
+            for (let tubeNum = 0; tubeNum < currTubes.length; tubeNum++) {
+                const fullTubeId = currTubes[tubeNum];
+                const tubeTypeStringArr = fullTubeId.split(' ');
+                const tubeType = translateNumToType[tubeTypeStringArr[1]] ? translateNumToType[tubeTypeStringArr[1]] : 'N/A';
+                const rowCount = shippingBoxContentsTable.rows.length;
+                const row = shippingBoxContentsTable.insertRow(rowCount);
+                
+                populateShippingBoxContentsRows(bagNum, tubeNum, row, currBagId, fullTubeId, tubeType);
+                if (tubeNum == 0) {
+                    const currDeleteButton = row.cells[3].querySelector(".delButton");
+                    handleRemoveBagButton(currDeleteButton, currTubes, currBoxId);
+                }
+            }
+        }
+    } else {
+      // Clear table if no list is found
+      const toInsertTable = document.getElementById('currTubeTable')
+      toInsertTable.innerHTML = renderViewBoxContentsTableHeader();
+    }
+}
+
+// Remove a bag from a box
+const handleRemoveBagButton = (currDeleteButton, currTubes, currBoxId) => {
+    currDeleteButton.addEventListener("click", async e => {
+        showAnimation();
+        const row = e.target.closest('tr');
+        const currBagId = row.cells[0].innerText;
+        const bagsToRemove = currBagId === "unlabelled" ? currTubes : [currBagId];        
+        const removeBagResponse = await removeBag(currBoxId, bagsToRemove);
+        hideAnimation();
+
+        if (removeBagResponse.code === 200) {
+            updateShippingStateRemoveBagFromBox(currBoxId, bagsToRemove);
+            startShipping(appState.getState().userName, true, currBoxId);
+        } else {
+            showNotifications({ title: 'Error removing bag', body: 'We experienced an error removing this bag. Please try again.' });
+        }
+    });
+}
+
+// Calculate the highest existing boxId and ++ for the new box.
+// Create the box and add it to firestore. On success, add it to the relevant state objects (allBoxesList and boxesByLocationList, and detailedProviderBoxes).
+export const addNewBox = async () => {    
+    const siteLocation = document.getElementById('selectLocationList').value;
+    const siteLocationConversion = siteSpecificLocationToConceptId[siteLocation];
+    const siteCode = siteSpecificLocation[siteLocation]["siteCode"];
+    
+    const boxList = appState.getState().allBoxesList;
+    
+    const { largestBoxIndex, largestBoxIndexAtLocation } = findLargestBoxData(boxList, siteLocation);
+    const shouldUpdateBoxModal = largestBoxIndexAtLocation !== -1 && Object.keys(boxList[largestBoxIndexAtLocation]['bags']).length !== 0;
+    
+    let largestBoxId;
+    if (shouldUpdateBoxModal) largestBoxId = boxList[largestBoxIndex][conceptIds.shippingBoxId];
+    else largestBoxId = largestBoxIndex !== -1 ? boxList[largestBoxIndex][conceptIds.shippingBoxId] : 'Box0';
+
+    if (largestBoxIndexAtLocation == -1 || shouldUpdateBoxModal) {
+        const boxToAdd = await createNewBox(boxList, siteLocationConversion, siteCode, largestBoxId, shouldUpdateBoxModal);
+        if (!boxToAdd) return false;
+        updateShippingStateCreateBox(boxToAdd);
+        return true;
+    } else {
+        return false;
+    }
+}
+
+const createNewBox = async (boxList, pageLocationConversion, siteCode, largestBoxId, updateBoxModal) => {
+    const newBoxNum = parseInt(largestBoxId.substring(3)) + 1;
+    const newBoxId = 'Box' + newBoxNum.toString();
+    const boxToAdd = {
+        'bags': {},
+        [conceptIds.shippingBoxId]: newBoxId,
+        [conceptIds.shippingLocation]: pageLocationConversion,
+        [conceptIds.siteCode]: siteCode,
+        [conceptIds.submitShipmentFlag]: conceptIds.no
     };
 
-    const element = document.getElementById(elemId);
-    if (element) {
-        element.removeEventListener("change", selectionChangeHandler);
-        element.addEventListener("change", selectionChangeHandler);
+    try {
+        await addBox(boxToAdd);
+    } catch (e) {
+        console.error('Error adding box', e);
+        return null;
     }
+    
+    boxList.push(boxToAdd);
+    if (updateBoxModal) document.getElementById('shippingModalChooseBox').setAttribute('data-new-box', newBoxId);
+
+    return boxToAdd;
+}
+
+// Find the largest box among all boxes and the largest box at the specified location.
+const findLargestBoxData = (boxList, siteLocation) => {
+    let largestBoxId = 0; 
+    let largestBoxIndex = -1;
+    let largestLocationBoxId = 0; 
+    let largestBoxIndexAtLocation = -1;
+
+    for (let i = 0; i < boxList.length; i++) {
+        const currBoxNum = parseInt(boxList[i][conceptIds.shippingBoxId].substring(3));
+        const currLocation = conceptIdToSiteSpecificLocation[boxList[i][conceptIds.shippingLocation]];
+
+        if (currBoxNum > largestBoxId) {
+            largestBoxId = currBoxNum;
+            largestBoxIndex = i;
+        }
+
+        if (currLocation == siteLocation && currBoxNum > largestLocationBoxId) {
+            largestLocationBoxId = currBoxNum;
+            largestBoxIndexAtLocation = i;
+        }
+    }
+
+    return { largestBoxIndex, largestBoxIndexAtLocation };
+}
+
+
+
+
+export const generateBoxManifest = (currBox) => {
+    const currInstitute = currBox.boxData.siteAcronym;
+    const currLocation = locationConceptIDToLocationMap[currBox.boxData[conceptIds.shippingLocation]]["siteSpecificLocation"];
+    const currContactInfo = locationConceptIDToLocationMap[currBox.boxData[conceptIds.shippingLocation]]["contactInfo"];
+
+    removeActiveClass('navbar-btn', 'active');
+    const navBarBoxManifestBtn = document.getElementById('navBarBoxManifest');
+    navBarBoxManifestBtn.classList.add('active');
+    document.getElementById('contentBody').innerHTML = renderBoxManifestTemplate(currInstitute, currLocation);
+
+    populateBoxManifestHeader(currBox, currContactInfo);
+    populateBoxManifestTable(currBox);
+    document.getElementById('printBox').addEventListener('click', e => {
+        window.print();
+    });
+
+    addEventReturnToPackaging();
+}
+
+export const populateBoxManifestHeader = (currBox, currContactInfo) => {
+    if(!currBox) return;
+
+    const currKeys = Object.keys(currBox).filter(key => key !== 'boxData' && key !== 'undefined');
+    const numBags = currKeys.length;
+    const numTubes = currKeys.reduce((acc, bagKey) => acc + currBox[bagKey]['arrElements'].length, 0);
+
+    const boxId = currBox.boxData[conceptIds.shippingBoxId];
+    const boxStartedTimestamp = formatTimestamp(currBox.boxData[conceptIds.firstBagAddedToBoxTimestamp]);
+    const boxLastModifiedTimestamp = formatTimestamp(currBox.boxData[conceptIds.shippingShipDateModify]);
+
+    renderBoxManifestHeader(boxId, boxStartedTimestamp, boxLastModifiedTimestamp, numBags, numTubes, currContactInfo);
+}
+
+// Render the box manifest header.
+// CreateParent divs, create list of data to be appended, create the new elements, append.
+const renderBoxManifestHeader = (boxId, boxStartedTimestamp, boxLastModifiedTimestamp, numBags, numTubes, currContactInfo) => {
+    const boxManifestCol1 = document.getElementById('boxManifestCol1');
+    const boxManifestCol3 = document.getElementById('boxManifestCol3');
+    
+    const div1 = document.createElement("div");
+    const div3 = document.createElement("div");
+    
+    const dataCol1 = [
+        { text: `${boxId} Manifest`, style: { fontWeight: '700', fontSize: '1.5rem' } },
+        { text: `Date Started: ${boxStartedTimestamp}` },
+        { text: `Last Modified: ${boxLastModifiedTimestamp}` },
+        { text: displayContactInformation(currContactInfo), isHTML: true }
+    ];
+    
+    const dataCol3 = [
+        { text: `Number of Sleeves/Bags: ${numBags}` },
+        { text: `Number of Specimens:  ${numTubes}` }
+    ];
+    
+    const createElements = (data, parent) => {
+        for (const item of data) {
+            const newP = document.createElement("p");
+            newP.innerHTML = item.text;
+    
+            if (item.style) {
+                Object.assign(newP.style, item.style);
+            }
+    
+            if (item.isHTML) {
+                const newDiv = document.createElement("div");
+                newDiv.innerHTML = newP.outerHTML;
+                parent.appendChild(newDiv);
+            } else {
+                parent.appendChild(newP);
+            }
+        }
+    }
+    
+    createElements(dataCol1, div1);
+    createElements(dataCol3, div3);
+    
+    boxManifestCol1.appendChild(div1);
+    boxManifestCol3.appendChild(div3);
+}
+
+// Get all ids from the available collections table and hidden orphan table.
+export const buildSpecimenDataInModal = (masterSpecimenId) =>{
+    const shippingTable = document.getElementById('specimenList');
+    const orphanTable = document.getElementById('orphansList');
+    
+    let foundInOrphan = false;
+    let biospecimensList = []
+    let tableIndex = -1;
+
+    for (let i = 1; i < shippingTable.rows.length; i++) {
+        const currRow = shippingTable.rows[i];
+        if (currRow.cells[0] !== undefined && currRow.cells[0].innerText == masterSpecimenId.toUpperCase()) {
+            tableIndex = i;
+            biospecimensList = JSON.parse(currRow.cells[2].innerText)
+        }
+    }
+
+    for (let i = 1; i < orphanTable.rows.length; i++) {
+        const currRow = orphanTable.rows[i];
+        if (currRow.cells[0] !== undefined && currRow.cells[0].innerText == masterSpecimenId.toUpperCase()) {
+            tableIndex = i;
+            const currTubeNum = currRow.cells[0].innerText.split(' ')[1];
+            biospecimensList = [currTubeNum];
+            foundInOrphan = true;
+        }
+    }
+
+    return { foundInOrphan, biospecimensList, tableIndex };
+}
+
+export const createShippingModalBody = (biospecimensList, masterBiospecimenId, isOrphan) => {
+    const boxList = appState.getState().boxesByLocationList;
+    let boxIdAndBagsObj = {};
+    for (let i = 0; i < boxList.length; i++) {
+        const box = boxList[i];
+        boxIdAndBagsObj[box[conceptIds.shippingBoxId]] = box['bags'];
+    }
+    
+    const tubeTable = document.createElement('table');
+    const splitTubeIdArray = masterBiospecimenId.split(/\s+/); /* Ex. ['CXA000133', '0008']*/
+    let isBagEmpty = true;
+    
+    biospecimensList = sortBiospecimensList(biospecimensList);
+    for (const specimenId of biospecimensList) {
+        if (shouldAddModalRow(isOrphan, splitTubeIdArray, specimenId)) {
+            isBagEmpty = addRowToModalTable(isBagEmpty, tubeTable, splitTubeIdArray, specimenId);
+        }
+    }
+
+    renderShippingModalBody(tubeTable.innerHTML);
+    populateModalSelect(boxIdAndBagsObj);
+
+    if (isBagEmpty) {
+        showNotifications({ title: 'Not found', body: 'The participant with entered search criteria not found!' }, true);
+        document.getElementById('shippingCloseButton').click();
+        hideAnimation();
+    }
+}
+
+const shouldAddModalRow = (isOrphan, splitTubeIdArray, tubeId) => {
+    if (isOrphan) return true;
+    if (splitTubeIdArray.length >= 2 && splitTubeIdArray[1] == '0008') {
+        //look for all non-mouthwash (0007)
+        return tubeId != '0007' && tubeId != '0008';
+    } else {
+        return tubeId == '0007' && tubeId != '0009';
+    }
+}
+
+const addRowToModalTable = (isBagEmpty, tubeTable, splitTubeIdArray, tubeId) => {
+    isBagEmpty = false;
+    const rowCount = tubeTable.rows.length;
+    const row = tubeTable.insertRow(rowCount);
+    const tubeType = translateNumToType.hasOwnProperty(tubeId) ? translateNumToType[tubeId] : 'N/A';
+
+    row.insertCell(0).innerHTML = `${splitTubeIdArray[0]} ${tubeId}`;
+    row.insertCell(1).innerHTML = tubeType;
+    row.insertCell(2).innerHTML = '<input type="checkbox" class="samplePresentCheckbox" style="transform: scale(2); display:block; margin:0 auto;" checked>';
+    row.cells[2].style.verticalAlign = "middle";
+
+    const checkboxEl = row.cells[2].firstChild;
+    checkboxEl.setAttribute("data-full-specimen-id", `${splitTubeIdArray[0]} ${tubeId}`);
+    checkboxEl.addEventListener("click", e => {
+        e.target.toggleAttribute("checked");
+    });
+
+    return isBagEmpty;
+}
+
+export const updateBoxListModalUIValue = () => {
+    const shippingModalChooseBox = document.getElementById('shippingModalChooseBox');
+    const boxId = shippingModalChooseBox.value;
+    document.getElementById('selectBoxList').value = boxId;
+    
+    return boxId;
+}
+
+export const processCheckedModalElements = (boxIdAndBagsObj, bagId, boxId, isOrphan, tableIndex) => {
+    const allCheckboxEle = document.querySelectorAll(".samplePresentCheckbox");
+    const [firstName = '', lastName = ''] = appState.getState().userName.split(/\s+/);
+    const checkedEleList = Array.from(allCheckboxEle).filter(ele => ele.checked);
+    const tubesToDelete = [];
+
+    if (isOrphan) bagId = 'unlabelled';
+
+    for (const checkedEle of checkedEleList) {
+        const specimenIdToAdd = checkedEle.getAttribute("data-full-specimen-id"); // data-full-specimen-id (Ex. "CXA444444 0007")
+        const [collectionId, tubeId] = specimenIdToAdd.split(/\s+/);
+        tubesToDelete.push(tubeId);
+
+        if (!isOrphan) {
+            bagId = assignBagId(tubeId, collectionId);
+        }
+
+        if (boxIdAndBagsObj.hasOwnProperty(boxId)) {
+            if (boxIdAndBagsObj[boxId].hasOwnProperty(bagId)) {
+                boxIdAndBagsObj[boxId][bagId]['arrElements'].push(specimenIdToAdd);
+            } else {
+                boxIdAndBagsObj[boxId][bagId] = {
+                    'arrElements': [specimenIdToAdd],
+                    [conceptIds.scannedByFirstName]: firstName,
+                    [conceptIds.scannedByLastName]: lastName
+                };
+            }
+        } else {
+            boxIdAndBagsObj[boxId] = {}
+            boxIdAndBagsObj[boxId][bagId] = {
+                'arrElements': [specimenIdToAdd],
+                [conceptIds.scannedByFirstName]: firstName,
+                [conceptIds.scannedByLastName]: lastName
+            };
+        }
+    }
+    handleAvailableCollectionsTableRows(tableIndex, tubesToDelete);
+    
+    return boxIdAndBagsObj;
+}
+
+export const prepareBoxToUpdate = (boxId, boxList, boxIdAndBagsObj, locations) => {
+    const currTime = new Date().toISOString();
+    const foundBox = boxList.find(box => box[conceptIds.shippingBoxId] == boxId) || {};
+
+    return {
+        ...foundBox,
+        'bags': boxIdAndBagsObj[boxId],
+        [conceptIds.shippingShipDateModify]: currTime,
+        [conceptIds.shippingBoxId]: boxId,
+        [conceptIds.shippingLocation]: locations[boxId],
+        [conceptIds.siteCode]: siteSpecificLocation[conceptIdToSiteSpecificLocation[locations[boxId]]].siteCode,
+        [conceptIds.firstBagAddedToBoxTimestamp]: foundBox[conceptIds.firstBagAddedToBoxTimestamp]
+            ? foundBox[conceptIds.firstBagAddedToBoxTimestamp]
+            : currTime,
+    };
+}
+
+const handleAvailableCollectionsTableRows = (tableIndex, tubesToDelete) => {
+    const availableCollectionsTable = document.getElementById('specimenList');
+    
+    // handle an orphan tube scanned if currArr is undefined  
+    const currArr = availableCollectionsTable?.rows[tableIndex]?.cells[2]?.innerText;
+    if(currArr != undefined) {
+        const parseCurrArr = JSON.parse(availableCollectionsTable.rows[tableIndex].cells[2].innerText);
+        for (let i = 0; i < tubesToDelete.length; i++) {
+            parseCurrArr.splice(parseCurrArr.indexOf(tubesToDelete[i]), 1);
+        }
+        if (parseCurrArr.length == 0) {
+            availableCollectionsTable.deleteRow(tableIndex);
+        } else {
+            availableCollectionsTable.rows[tableIndex].cells[2].innerText = JSON.stringify(parseCurrArr);
+            availableCollectionsTable.rows[tableIndex].cells[1].innerText = parseCurrArr.length;
+        }
+    }
+}
+
+const assignBagId = (tubeId, collectionId) => {
+    if (tubeId === '0007') {
+        return collectionId + ' 0009';
+    } else {
+        return collectionId + ' 0008';
+    }
+}
+
+export const populateModalSelect = (detailedProvierBoxes) => {
+    const boxSelectEle = document.getElementById('shippingModalChooseBox');
+    const selectedBoxId = boxSelectEle.getAttribute('data-new-box') || document.getElementById('selectBoxList').value;
+    const addToBoxButton =  document.getElementById('addToBoxButton');
+    
+    addToBoxButton.removeAttribute("disabled")
+    
+    const boxIds = Object.keys(detailedProvierBoxes).sort(compareBoxIds);
+    const options = boxIds.map(boxId => `<option>${boxId}</option>`).join(''); 
+
+    if (options == '') {
+        addToBoxButton.setAttribute('disabled', 'true');
+    }
+
+    boxSelectEle.innerHTML = options;
+    boxSelectEle.value = selectedBoxId;
+}
+
+export const populateTempSelect = (boxes) => {
+    const boxDiv = document.getElementById("tempCheckList");
+    boxDiv.style.display = "block";
+    boxDiv.style.marginTop = "10px";
+    boxDiv.innerHTML = `<p>Select the box that contains the temperature monitor</p>
+        <select name="tempBox" id="tempBox">
+        <option disabled value> -- select a box -- </option>
+        </select>`;
+
+    const toPopulate = document.getElementById('tempBox')
+
+    for (let i = 0; i < boxes.length; i++) {
+        
+        var opt = document.createElement("option");
+        opt.value = boxes[i];
+        opt.innerHTML = boxes[i];
+        if(i === 0){
+            opt.selected = true;
+        }
+        // then append it to the select element
+        toPopulate.appendChild(opt);
+    }
+}
+
+export const formatTimestamp = (timestamp) => {
+    const newDate = timestamp ? new Date(timestamp) : new Date();
+    const ampm = newDate.getHours() >= 12 ? 'PM' : 'AM';
+    const minutesTag = newDate.getMinutes() < 10 ? `0${newDate.getMinutes()}` : newDate.getMinutes();
+    return `${newDate.getMonth() + 1}/${newDate.getDate()}/${newDate.getFullYear()} ${(newDate.getHours() + 11) % 12 + 1}:${minutesTag} ${ampm}`;
+};
+
+// get available locations from firestore on the first load
+// save those locations in state and load from state on subsequent loads - if (availableLocations)
+const populateSelectLocationList = async (availableLocations, loadFromState) => {
+    const locationSelection = JSON.parse(localStorage.getItem('selections'))?.shipping_location;
+    const selectEle = document.getElementById('selectLocationList');
+    
+    if (!loadFromState) {
+        availableLocations = await getLocationsInstitute();
+    }
+
+    const defaultOption = `<option value="none" ${locationSelection === 'none' ? 'selected="selected"' : ""}>Select Shipping Location</option>`;    
+    const locationOptions = availableLocations.map(location => 
+        `<option ${locationSelection === location ? 'selected="selected"' : ""} value="${location}">${location}</option>`
+    ).join('');
+
+    selectEle.innerHTML = defaultOption + locationOptions;
+
+    return availableLocations;
+}
+
+const populateBoxManifestTable = (currBox) => {
+    const boxManifestTable = document.getElementById('boxManifestTable');
+    const bagList = Object.keys(currBox).filter(key => key !== 'boxData' && key !== 'undefined').sort(sortSpecimenKeys);
+
+    bagList.forEach((bagKey, i) => {
+        const tubesList = currBox[bagKey].arrElements;
+
+        for (let i = 0; i < tubesList.length; i++) {
+            const tubeDetail = currBox[bagKey].specimenDetails[tubesList[i]];
+            const currRow = boxManifestTable.insertRow(i + 1);
+            const tubeId = tubesList[i].split(' ');
+            const tubeTypeAndColor = translateNumToType[tubeId[1]] ? translateNumToType[tubeId[1]] : 'N/A';
+
+            currRow.insertCell(0).innerHTML = i === 0 ? bagKey : '';
+            currRow.insertCell(1).innerHTML = tubesList[i];
+            currRow.insertCell(2).innerHTML = tubeTypeAndColor;
+
+            addDeviationTypeCommentsContent(tubeDetail, currRow, i);
+        };
+    });
+}
+
+const sortSpecimenKeys = (a, b) => {
+    const numA = parseInt(a.split(' ')[0].substring(3));
+    const numB = parseInt(b.split(' ')[0].substring(3));
+    return numB - numA;
+}
+
+/**
+ * 
+ * @param {string[]} boxIdArray
+ * @param {string} userName
+ * @param {boolean} isTempMonitorIncluded
+ * @param {*} currShippingLocationNumber
+ */
+export const generateShippingManifest = (boxIdArray, userName, isTempMonitorIncluded, currShippingLocationNumber) => {
+    const boxArray = appState.getState().allBoxesList;
+    const boxIdAndBagsObj = {};
+    const locations = {};
+    let site = '';
+
+    for (const box of boxArray) {
+        const boxId= box[conceptIds.shippingBoxId]
+        boxIdAndBagsObj[boxId] = box['bags']
+        locations[boxId] = box[conceptIds.shippingLocation];
+        site = box['siteAcronym'];
+    }
+    
+    let boxIdAndBagsObjToDisplay = {};
+    let location = '';
+
+    for (const boxId of boxIdArray) {
+        boxIdAndBagsObjToDisplay[boxId] = boxIdAndBagsObj[boxId];
+        location = locations[boxId];
+    }
+
+    removeActiveClass('navbar-btn', 'active')
+    const navBarBtn = document.getElementById('navBarReviewShipmentContents');
+    navBarBtn.classList.add('active');
+    document.getElementById('contentBody').innerHTML = renderShippingManifestTemplate();//template;
+
+    if(isTempMonitorIncluded){
+        populateTempSelect(boxIdArray);
+    }
+
+    document.getElementById('shippingHiddenTable').innerHTML = JSON.stringify(boxIdAndBagsObj);
+    
+    populateShippingManifestHeader(userName, site, currShippingLocationNumber); // populate shipping header via site specfiic location selected from shipping page
+    populateShippingManifestTable(boxIdAndBagsObjToDisplay);
+    addEventNavBarShipment("navBarShippingDash", userName);
+    populateTempCheck();
+    const btn = document.getElementById('assignTrackingNumberPage'); // assignTracking
+    addEventShipPrintManifest('printBox')
+    addEventReturnToPackaging();    
+
+    document.getElementById('assignTrackingNumberPage').addEventListener('click', e => {
+        e.stopPropagation();
+        if(btn.classList.contains('active')) return;
+
+        const tempBoxElement = document.getElementById('tempBox');
+
+        if (isTempMonitorIncluded && tempBoxElement.value === '') {
+            showNotifications({title: 'Missing field!', body: 'Please enter the box where the temperature monitor is being stored.'}, true)
+            return;
+        }
+
+        let boxWithTempMonitor = '';
+        if( isTempMonitorIncluded){
+            boxWithTempMonitor = tempBoxElement.value;
+        }
+
+        shipmentTracking(boxIdAndBagsObjToDisplay, userName, boxWithTempMonitor);
+    });
+}
+
+export const populateShippingManifestHeader = (userName, siteAcronym, currShippingLocationNumber) => {
+    const currContactInfo = locationConceptIDToLocationMap[currShippingLocationNumber]["contactInfo"][siteAcronym];
+    const siteSpecificLocation = locationConceptIDToLocationMap[currShippingLocationNumber]["siteSpecificLocation"];
+
+    const currentDateTime = formatTimestamp();
+    const dataCol1 = [
+        { text: "Shipment Manifest", style: { fontWeight: '700', fontSize: '1.5rem' } },
+        { text: `Current Date/Time: ${currentDateTime}` },
+        { text: `Sender: ${userName}` },
+        { text: displayContactInformation(currContactInfo), isHTML: true }
+    ];
+    
+    const dataCol3 = [
+        { text: `Site: ${siteAcronym}` },
+        { text: `Location: ${siteSpecificLocation}` }
+    ];
+
+    buildShippingManifestHeader(dataCol1, dataCol3);
+}
+
+const buildShippingManifestHeader = (dataCol1, dataCol3) => {
+    const boxManifestCol1 = document.getElementById('boxManifestCol1');
+    const boxManifestCol3 = document.getElementById('boxManifestCol3');
+    
+    const div1 = document.createElement("div");
+    const div3 = document.createElement("div");
+    
+    const createElements = (data, parent) => {
+        for (const item of data) {
+            const newP = document.createElement("p");
+            newP.innerHTML = item.text;
+
+            if (item.style) {
+                Object.assign(newP.style, item.style);
+            }
+    
+            if (item.isHTML) {
+                const newDiv = document.createElement("div");
+                newDiv.innerHTML = newP.outerHTML;
+                parent.appendChild(newDiv);
+            } else {
+                parent.appendChild(newP);
+            }
+        }
+    }
+    
+    createElements(dataCol1, div1);
+    createElements(dataCol3, div3);
+    
+    boxManifestCol1.appendChild(div1);
+    boxManifestCol3.appendChild(div3);
+}
+
+export const populateShippingManifestTable = (boxIdAndBagsObj) => {
+    const table = document.getElementById("shippingManifestTable");
+    const boxIdArray = Object.keys(boxIdAndBagsObj).sort(compareBoxIds);
+    let greyIndex = 0;
+    let firstSpecimenInBox = true;
+
+    boxIdArray.forEach((currBoxId) => {
+        const specimens = Object.keys(boxIdAndBagsObj[currBoxId]);        
+        specimens.forEach((specimen) => {
+            const tubes = boxIdAndBagsObj[currBoxId][specimen]['arrElements'];
+
+            tubes.forEach((currTube, tubeIndex) => {
+                const specimenData = boxIdAndBagsObj[currBoxId][specimen];
+                const fullScannerName = tubeIndex === 0 
+                    ? `${specimenData[conceptIds.scannedByFirstName] || ''} ${specimenData[conceptIds.scannedByLastName] || ''}`.trim()
+                    : '';
+                const currRow = table.insertRow(-1);
+                currRow.insertCell(0).innerHTML = firstSpecimenInBox && tubeIndex === 0 ? currBoxId : '';
+                currRow.insertCell(1).innerHTML = tubeIndex === 0 ? specimen : '';
+                currRow.insertCell(2).innerHTML = currTube;
+                currRow.insertCell(3).innerHTML = fullScannerName;
+
+                if (greyIndex % 2 === 0) currRow.style['background-color'] = "lightgrey";
+                if (firstSpecimenInBox && tubeIndex === 0) firstSpecimenInBox = false;
+            });
+
+            greyIndex += 1;
+        });
+
+        firstSpecimenInBox = true;
+    });
+}
+
+export const shipmentTracking = async (boxIdAndBagsObj, userName, boxWithTempMonitor) => {
+    showAnimation();
+
+    if(document.getElementById('navBarParticipantCheckIn')) document.getElementById('navBarParticipantCheckIn').classList.add('disabled');
+    let template = `
+        <div id="shippingHiddenTable" style="display:none">
+        {}
+        </div>
+        <div class="row" style="margin-top:40px;">
+            <div class="col-lg">
+                <label for="courierSelect" style="font-size:1.4rem; margin-bottom:1rem;">Choose Shipment Courier</label>
+                <select name="courier" id="courierSelect" style="padding:.2rem; display:block;">
+                </select>
+            </div>
+        </div>
+        <div class="row" style="margin-top:40px">
+            <div class="col-lg">
+                <p style="margin:0; font-size:1.4rem;">Enter Shipment Tracking Numbers:</p>
+                <div class="col-lg" id="forTrackingNumbers">
+                    
+                </div>
+            </div>
+        </div>
+        <div class="row" style="margin-top:100px">
+            <div style="float: left;width: 33%;" id="boxManifestCol1">
+                <button type="button" class="btn btn-primary" data-dismiss="modal" id="returnToPackaging">Return to Packaging</button>
+            </div>
+            <div style="float: left;width: 33%;" id="boxManifestCol2">
+                
+            </div>
+            <div style="float:left;width: 33%;" id="boxManifestCol3">
+                <button type="button" class="btn btn-primary" data-dismiss="modal" id="saveTracking" style="margin-right:.5rem;">Save</button>
+                <button type="button" class="btn btn-primary" data-dismiss="modal" id="completeTracking">Save and Continue</button>
+            </div>
+        </div>
+
+    `;
+
+    removeActiveClass('navbar-btn', 'active')
+    document.getElementById('contentHeader').innerHTML = `<h2>Connect for Cancer Prevention Study</h2></br>` + shippingNavBar();
+    const navBarBtn = document.getElementById('navBarShipmentTracking');
+    navBarBtn.classList.add('active');
+    document.getElementById('contentBody').innerHTML = template;
+    await populateCourierBox();
+    addEventNavBarShipment("returnToPackaging", userName);
+    if(Object.keys(boxIdAndBagsObj).length > 0){
+        document.getElementById('shippingHiddenTable').innerText = JSON.stringify(boxIdAndBagsObj)
+    }
+    addEventNavBarShipment("navBarShippingDash", userName);
+    addEventReturnToReviewShipmentContents('navBarReviewShipmentContents', boxIdAndBagsObj, userName, boxWithTempMonitor)
+    await populateTrackingQuery(boxIdAndBagsObj);
+    addEventTrimTrackingNums()
+    addEventTrackingNumberScanAutoFocus()
+    addEventPreventTrackingConfirmPaste()
+    addEventCheckValidTrackInputs(boxIdAndBagsObj)
+    addEventSaveButton(boxIdAndBagsObj);
+    addEventCompleteButton(boxIdAndBagsObj, userName, boxWithTempMonitor);
+    hideAnimation();
+}
+
+export const finalShipmentTracking = (boxIdAndBagsObj, userName, boxWithTempMonitor, shipmentCourier) => {
+    if(document.getElementById('navBarParticipantCheckIn')) document.getElementById('navBarParticipantCheckIn').classList.add('disabled');
+    
+    removeActiveClass('navbar-btn', 'active')
+    document.getElementById('contentHeader').innerHTML = `<h2 >Connect for Cancer Prevention Study</h2></br>` + shippingNavBar();
+    const navBarBtn = document.getElementById('navBarFinalizeShipment');
+    navBarBtn.classList.add('active');
+    document.getElementById('contentBody').innerHTML = renderFinalShipmentTrackingTemplate(shipmentCourier, userName);
+    
+    addEventNavBarShipment("navBarShippingDash", userName);
+    addEventNavBarTracking("returnToTracking", userName, boxIdAndBagsObj, boxWithTempMonitor)
+    addEventNavBarTracking("navBarFinalizeShipment", userName, boxIdAndBagsObj, boxWithTempMonitor)
+    if(Object.keys(boxIdAndBagsObj).length > 0){
+        document.getElementById('shippingHiddenTable').innerText = JSON.stringify(boxIdAndBagsObj)
+    }
+    populateFinalCheck(boxIdAndBagsObj);
+    addEventReturnToReviewShipmentContents('navBarReviewShipmentContents', boxIdAndBagsObj, userName)
+    addEventCompleteShippingButton(boxIdAndBagsObj, userName, boxWithTempMonitor, shipmentCourier);
+    addEventBackToSearch('navBarShippingDash');
 }
 
 const renderShippingHiddenTable = () => {
@@ -172,7 +1034,7 @@ const renderShippingLocationSelector = () => {
 
 const renderScanOrEnterSpecimenId = () => {
     return `
-    <div class="row" style="margin:0;">
+    <div class="row" style="margin:0; margin-bottom: 1.5rem;">
         <div class="col-lg" style="padding: 0;margin: 0;">
             <div class="row form-row" style="padding-left:0px;margin:0;">
                 <form id="addSpecimenForm" method="POST" style="width:100%;">
@@ -212,7 +1074,6 @@ const renderCollectionsContentsAndBoxes = () => {
     `;
 }
 
-//TODO orphan panel is always hidden
 const renderAvailableCollections = () => {
     return `
         <div class="col-5">
@@ -260,7 +1121,7 @@ const renderTempTubeReminder = () => {
 
 const renderSelectBoxes = () => {
     return `
-        <div class="row" style="margin-bottom:.5rem">
+        <div class="row" style="margin-bottom:1rem">
             <div class="col-9 no-gutters">
                 <h4 style="text-align:start;">Select one or more boxes to ship</h4>
             </div>
@@ -312,89 +1173,6 @@ const renderSpecimenVerificationModal = () => {
     `;
 }
 
-//TODO: move boxManifest items to separate file
-export const generateBoxManifest = (currBox) => {
-    const currInstitute = currBox.boxData.siteAcronym;
-    const currLocation = locationConceptIDToLocationMap[currBox.boxData[conceptIds.shippingLocation]]["siteSpecificLocation"];
-    const currContactInfo = locationConceptIDToLocationMap[currBox.boxData[conceptIds.shippingLocation]]["contactInfo"];
-
-    removeActiveClass('navbar-btn', 'active');
-    const navBarBoxManifestBtn = document.getElementById('navBarBoxManifest');
-    navBarBoxManifestBtn.classList.add('active');
-    document.getElementById('contentBody').innerHTML = renderBoxManifestTemplate(currInstitute, currLocation);
-
-    populateBoxManifestHeader(currBox, currContactInfo);
-    populateBoxManifestTable(currBox);
-    document.getElementById('printBox').addEventListener('click', e => {
-        window.print();
-    });
-
-    addEventReturnToPackaging("returnToPackaging", appState.getState().userName);
-}
-
-export const populateBoxManifestHeader = (currBox, currContactInfo) => {
-    if(!currBox) return;
-
-    const currKeys = Object.keys(currBox).filter(key => key !== 'boxData' && key !== 'undefined');
-    const numBags = currKeys.length;
-    const numTubes = currKeys.reduce((acc, bagKey) => acc + currBox[bagKey]['arrElements'].length, 0);
-
-    const boxId = currBox.boxData[conceptIds.shippingBoxId];
-    const boxStartedTimestamp = formatBoxTimestamp(currBox.boxData[conceptIds.firstBagAddedToBoxTimestamp]);
-    const boxLastModifiedTimestamp = formatBoxTimestamp(currBox.boxData[conceptIds.shippingShipDateModify]);
-
-    renderBoxManifestHeader(boxId, boxStartedTimestamp, boxLastModifiedTimestamp, numBags, numTubes, currContactInfo);
-}
-
-const renderBoxManifestHeader = (boxId, boxStartedTimestamp, boxLastModifiedTimestamp, numBags, numTubes, currContactInfo) => {
-    const boxManifestCol1 = document.getElementById('boxManifestCol1');
-    const boxManifestCol3 = document.getElementById('boxManifestCol3');
-    
-    // PParent divs
-    const div1 = document.createElement("div");
-    const div3 = document.createElement("div");
-    
-    // List of data to be appended
-    const dataCol1 = [
-        { text: `${boxId} Manifest`, style: { fontWeight: '700', fontSize: '1.5rem' } },
-        { text: `Date Started: ${boxStartedTimestamp}` },
-        { text: `Last Modified: ${boxLastModifiedTimestamp}` },
-        { text: displayContactInformation(currContactInfo), isHTML: true }
-    ];
-    
-    const dataCol3 = [
-        { text: `Number of Sleeves/Bags: ${numBags}` },
-        { text: `Number of Specimens:  ${numTubes}` }
-    ];
-    
-    // Function to create elements from data
-    const createElements = (data, parent) => {
-        for (const item of data) {
-            const newP = document.createElement("p");
-            newP.innerHTML = item.text;
-    
-            if (item.style) {
-                Object.assign(newP.style, item.style);
-            }
-    
-            if (item.isHTML) {
-                const newDiv = document.createElement("div");
-                newDiv.innerHTML = newP.outerHTML;
-                parent.appendChild(newDiv);
-            } else {
-                parent.appendChild(newP);
-            }
-        }
-    }
-    
-    createElements(dataCol1, div1);
-    createElements(dataCol3, div3);
-    
-    // Append parent divs to the target elements
-    boxManifestCol1.appendChild(div1);
-    boxManifestCol3.appendChild(div3);
-}
-
 const renderBoxManifestTemplate = (currInstitute, currLocation) => {
     return `
         </br>
@@ -442,36 +1220,77 @@ const renderBoxManifestTemplate = (currInstitute, currLocation) => {
     `;
 }
 
-/**
- * 
- * @param {string[]} boxIdArray
- * @param {string} userName
- * @param {boolean} isTempMonitorIncluded
- * @param {*} currShippingLocationNumber
- */
-export const shippingManifest = async (boxIdArray, userName, isTempMonitorIncluded, currShippingLocationNumber) => {
-    let response = await  getBoxes();
-    let boxArray = response.data;
-    let boxIdAndBagsObj = {};
-    let locations = {};
-    let site = '';
+export const renderShippingModalHeader = () => {
+    const header = document.getElementById('shippingModalHeader');
+        header.innerHTML = `<h5 class="modal-title">Specimen Verification</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close" id="shippingCloseButton">
+        <span aria-hidden="true">&times;</span>
+        </button>`;
+}
 
-    for (const box of boxArray) {
-        const boxId= box[conceptIds.shippingBoxId]
-        boxIdAndBagsObj[boxId] = box['bags']
-        locations[boxId] = box[conceptIds.shippingLocation];
-        site = box['siteAcronym'];
-    }
-    
-    let boxIdAndBagsObjToDisplay = {};
-    let location = ''
+export const renderShippingModalBody = (tubeTable) => {
+    document.getElementById('shippingModalBody').innerHTML = `
+    <table class="table" id="shippingModalTable">
+        <thead>
+            <tr>
+                <th>Full Specimen ID</th>
+                <th>Type/Color</th>
+                <th style="text-align:center;">Sample Present</th>
+            </tr>
+        </thead>
+        ${tubeTable}
+    </table>
+    `;
+}
 
-    for (const boxId of boxIdArray) {
-        boxIdAndBagsObjToDisplay[boxId] = boxIdAndBagsObj[boxId];
-        location = locations[boxId];
-    }
+export const buildPopulateSpecimensHeader = () => {
+    return `
+        <tr>
+            <th>Specimen Bag ID</th>
+            <th># Specimens in Bag</th>
+        </th>
+    `;
+}
 
-    let template = `
+export const renderBoxesToShipTableHeader = () => {
+    return `
+        <tr>
+            <th style="border-bottom:1px solid;">To Ship</th>
+            <th style="border-bottom:1px solid;">Started</th>
+            <th style="border-bottom:1px solid;">Last Modified</th>
+            <th style="border-bottom:1px solid;">Box Number</th>
+            <th style="border-bottom:1px solid;">Location</th>
+            <th style="border-bottom:1px solid;">Contents</th>
+            <th style="border-bottom:1px solid;text-align:center;"><p style="margin-bottom:0">View/Print Box Manifest</p><p style="margin-bottom:0">(to be included in shipment)</p></th>
+        </tr>
+    `;
+}
+
+export const renderBoxesToShipRow = (boxStartedTimestamp, boxLastModifiedTimestamp, boxId, boxLocation, numTubesInBox) => {
+    return `
+        <td><input type="checkbox" class="markForShipping" style="transform: scale(1.5); text-align= center;"></td>
+        <td>${boxStartedTimestamp}</td>
+        <td>${boxLastModifiedTimestamp}</td>
+        <td>${boxId}</td>
+        <td>${boxLocation}</td>
+        <td>${numTubesInBox} tubes</td>
+        <td><input type="button" style="display:block;margin:0 auto;" class="boxManifestButton" value="Box Manifest"/></td>
+    `;
+};
+
+export const renderViewBoxContentsTableHeader = () => {
+    return `
+        <tr>
+            <th style = "border-bottom:1px solid;">Specimen Bag ID</th>
+            <th style = "border-bottom:1px solid;">Full Specimen ID</th>
+            <th style = "border-bottom:1px solid;">Type/Color</th>
+            <th style = "border-bottom:1px solid;"></th>
+        </tr>
+    `;
+}
+
+export const renderShippingManifestTemplate = () => {
+    return `
         </br>
         <div id="shippingHiddenTable" style="display:none">
             <table>
@@ -512,149 +1331,11 @@ export const shippingManifest = async (boxIdArray, userName, isTempMonitorInclud
                 <button type="button" class="btn btn-primary" data-dismiss="modal" id="assignTrackingNumberPage">Continue to Assign Tracking Number</button>
             </div>
         </div>
-        
-        `;
-        /*
-        <div class="row" style="margin-top:100px">
-            <div style="float: left;width: 33%;" id="boxManifestCol1">
-                <button type="button" class="btn btn-primary" data-dismiss="modal" id="returnToPackaging">Return to Packaging</button>
-            </div>
-            <div style="float: left;width: 33%;">
-                <button type="button" class="btn btn-primary" data-dismiss="modal" id="printBox">Print Box Manifest</button>
-            </div>
-            <div style="float:left;width: 33%;" id="boxManifestCol3">
-                <button type="button" class="btn btn-primary" data-dismiss="modal" id="completePackaging">Packaging Complete</button>
-            </div>
-        </div>
-        `;*/
-    removeActiveClass('navbar-btn', 'active')
-    const navBarBtn = document.getElementById('navBarReviewShipmentContents');
-    navBarBtn.classList.add('active');
-    document.getElementById('contentBody').innerHTML = template;
-
-    if(isTempMonitorIncluded){
-        populateTempSelect(boxIdArray);
-    }
-
-    document.getElementById('shippingHiddenTable').innerHTML = JSON.stringify(boxIdAndBagsObj);
-    
-    
-    //document.getElementById('boxManifestTable').appendChild(result);
-    
-    populateShippingManifestHeader(userName, site, currShippingLocationNumber); // populate shipping header via site specfiic location selected from shipping page
-    populateShippingManifestBody(boxIdAndBagsObjToDisplay);
-    addEventNavBarShipment("navBarShippingDash", userName);
-    await populateTempCheck();
-    const btn = document.getElementById('assignTrackingNumberPage'); // assignTracking
-    addEventShipPrintManifest('printBox')
-    addEventNavBarShipment('returnToPackaging', userName);
-
-    
-
-    document.getElementById('assignTrackingNumberPage').addEventListener('click', e => {
-        e.stopPropagation();
-        if(btn.classList.contains('active')) return;
-
-        const tempBoxElement = document.getElementById('tempBox');
-
-        if (isTempMonitorIncluded && tempBoxElement.value === '') {
-            showNotifications({title: 'Missing field!', body: 'Please enter the box where the temperature monitor is being stored.'}, true)
-            return;
-        }
-
-        let boxWithTempMonitor = '';
-        if( isTempMonitorIncluded){
-            boxWithTempMonitor = tempBoxElement.value;
-        }
-
-        shipmentTracking(boxIdAndBagsObjToDisplay, userName, boxWithTempMonitor);
-    });
-}
-
-
-export const shipmentTracking = async (boxIdAndBagsObj, userName, boxWithTempMonitor) => {
-    showAnimation();
-
-    if(document.getElementById('navBarParticipantCheckIn')) document.getElementById('navBarParticipantCheckIn').classList.add('disabled');
-    //store a secret json that has all of the packed ones in it
-    //{"Box1":{specimenId:[allTubes], specimenId:[allTubes]}}
-    
-    /*
-    let hiddenJSON = {};
-    if(document.getElementById('shippingHiddenTable') != null){
-        hiddenJSON = JSON.parse(document.getElementById('shippingHiddenTable').innerText);
-    }
-*/
-    let template = `
-        <div id="shippingHiddenTable" style="display:none">
-        {}
-        </div>
-        <div class="row" style="margin-top:40px;">
-            <div class="col-lg">
-                <label for="courierSelect" style="font-size:1.4rem; margin-bottom:1rem;">Choose Shipment Courier</label>
-                <select name="courier" id="courierSelect" style="padding:.2rem; display:block;">
-                </select>
-            </div>
-        </div>
-        <div class="row" style="margin-top:40px">
-            <div class="col-lg">
-                <p style="margin:0; font-size:1.4rem;">Enter Shipment Tracking Numbers:</p>
-                <div class="col-lg" id="forTrackingNumbers">
-                    
-                </div>
-            </div>
-        </div>
-        <div class="row" style="margin-top:100px">
-            <div style="float: left;width: 33%;" id="boxManifestCol1">
-                <button type="button" class="btn btn-primary" data-dismiss="modal" id="returnToPackaging">Return to Packaging</button>
-            </div>
-            <div style="float: left;width: 33%;" id="boxManifestCol2">
-                
-            </div>
-            <div style="float:left;width: 33%;" id="boxManifestCol3">
-                <button type="button" class="btn btn-primary" data-dismiss="modal" id="saveTracking" style="margin-right:.5rem;">Save</button>
-                <button type="button" class="btn btn-primary" data-dismiss="modal" id="completeTracking">Save and Continue</button>
-            </div>
-        </div>
-
     `;
-    /*var x = document.getElementById("specimenList");
-    var option = document.createElement("option");
-    option.text = "Kiwi";
-    x.add(option);*/
-    
-    
-    removeActiveClass('navbar-btn', 'active')
-    document.getElementById('contentHeader').innerHTML = `<h2>Connect for Cancer Prevention Study</h2></br>` + shippingNavBar();
-    const navBarBtn = document.getElementById('navBarShipmentTracking');
-    navBarBtn.classList.add('active');
-    document.getElementById('contentBody').innerHTML = template;
-    await populateCourierBox();
-    addEventNavBarShipment("returnToPackaging", userName);
-    if(Object.keys(boxIdAndBagsObj).length > 0){
-        document.getElementById('shippingHiddenTable').innerText = JSON.stringify(boxIdAndBagsObj)
-    }
-    //addEventReturnToShippingManifest('returnToShipping', hiddenJSON, userName, tempCheckChecked)
-    addEventNavBarShipment("navBarShippingDash", userName);
-    addEventReturnToReviewShipmentContents('navBarReviewShipmentContents', boxIdAndBagsObj, userName, boxWithTempMonitor)
-    await populateTrackingQuery(boxIdAndBagsObj);
-    addEventTrimTrackingNums()
-    addEventTrackingNumberScanAutoFocus()
-    addEventPreventTrackingConfirmPaste()
-    addEventCheckValidTrackInputs(boxIdAndBagsObj)
-    addEventSaveButton(boxIdAndBagsObj);
-    addEventCompleteButton(boxIdAndBagsObj, userName, boxWithTempMonitor);
-    //addEventCompleteShippingButton(hiddenJSON);
-    //addEventBackToSearch('navBarShippingDash');
-    // addEventBarCodeScanner('masterSpecimenIdBarCodeBtn', 0, 9, 0);
-    hideAnimation();
-    //addEventSubmitAddBag();
 }
 
-export const finalShipmentTracking = (boxIdAndBagsObj, userName, boxWithTempMonitor, shipmentCourier) => {
-    if(document.getElementById('navBarParticipantCheckIn')) document.getElementById('navBarParticipantCheckIn').classList.add('disabled');
-
-    let template = `
+const renderFinalShipmentTrackingTemplate = (shipmentCourier, userName) => {
+    return `
         <div id="shippingHiddenTable" style="display:none">
         {}
         </div>
@@ -685,7 +1366,7 @@ export const finalShipmentTracking = (boxIdAndBagsObj, userName, boxWithTempMoni
                 <button type="button" class="btn btn-primary" data-dismiss="modal" id="returnToTracking">Back to Assign Tracking Information</button>
             </div>
             <div style="float: left;width: 33%;">
-                 
+                
             </div>
             <div style="float:left;width: 33%;" id="boxManifestCol3">
                 <button type="button" class="btn btn-primary"  data-toggle="modal" data-target="#finalizeModal" id="completeShippingButton">Finalize</button>
@@ -718,25 +1399,5 @@ export const finalShipmentTracking = (boxIdAndBagsObj, userName, boxWithTempMoni
                 </div>
             </div>
         </div>
-
-
     `;
-    
-    removeActiveClass('navbar-btn', 'active')
-    document.getElementById('contentHeader').innerHTML = `<h2 >Connect for Cancer Prevention Study</h2></br>` + shippingNavBar();
-    const navBarBtn = document.getElementById('navBarFinalizeShipment');
-    navBarBtn.classList.add('active');
-    document.getElementById('contentBody').innerHTML = template;
-    
- 
-    addEventNavBarShipment("navBarShippingDash", userName);
-    addEventNavBarTracking("returnToTracking", userName, boxIdAndBagsObj, boxWithTempMonitor)
-    addEventNavBarTracking("navBarFinalizeShipment", userName, boxIdAndBagsObj, boxWithTempMonitor)
-    if(Object.keys(boxIdAndBagsObj).length > 0){
-        document.getElementById('shippingHiddenTable').innerText = JSON.stringify(boxIdAndBagsObj)
-    }
-    populateFinalCheck(boxIdAndBagsObj);
-    addEventReturnToReviewShipmentContents('navBarReviewShipmentContents', boxIdAndBagsObj, userName)
-    addEventCompleteShippingButton(boxIdAndBagsObj, userName, boxWithTempMonitor, shipmentCourier);
-    addEventBackToSearch('navBarShippingDash');
 }
